+++ docs/conf.py
@@ -16,7 +16,7 @@ sys.path.insert(0, os.path.abspath('..'))
 
 # -- Project information -----------------------------------------------------
 
-project = 'temscript'
+project = 'pytemscript'
 copyright = '2012-2021, Tore Niermann; 2022, Grigory Sharov'
 author = 'Tore Niermann, Grigory Sharov'
 
@@ -41,7 +41,7 @@ language = 'en'
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ['_build', '../temscript/old', 'server.rst']
+exclude_patterns = ['_build', '../pytemscript/old', 'server.rst']
 
 # The name of the Pygments (syntax highlighting) style to use.
 pygments_style = 'sphinx'