+++ scripts/run_null_server.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-from temscript.old.server import run_server
+from pytemscript.old.server import run_server
 
 if __name__ == '__main__':
     run_server(['--null'])