+++ temscript/__init__.py
@@ -1,2 +1,0 @@
-from temscript.microscope import Microscope
-__version__ = '3.0a1'